We have DISCARDING!!!
When a player clicks a tile, I’ll discard the tile and put it in the
discard pile! :-)

I also put the [de]activate functions in an evaluate method that can be
called when needed.

Also, refreshHandUi now completely updates the current player’s hand
(this will be called at the end of each players’ turn)