Game flow has almost started! (read description)
I’ve added the performTurn method which will be responsible for the
bots taking turns. Basically, if a human player presses a tile to
discard, it will increment the currentPlayer and invoke performTurn.
PerformTurn then checks the currentPlayer, and evaluates the
corresponding hand and then performs logic. Once logic is performed,
then it increments currentPlayer and invokes itself but only if it is
bot 1 or 2. If it is bot 3, it doesn’t invoke performTurn because it
will then be the human’s turn so the onTouchListeners will be listening
for an action to which that will invoke performTurn which starts the
process again.